+++ crates/bcr-ebill-wasm/src/data/bill.rs
@@ -16,10 +16,7 @@ use serde::{Deserialize, Serialize};
 use tsify::Tsify;
 use wasm_bindgen::prelude::*;
 
-use super::{
-    FileWeb, FromWeb, IntoWeb, PostalAddressWeb, contact::ContactTypeWeb,
-    notification::NotificationWeb,
-};
+use super::{FileWeb, PostalAddressWeb, contact::ContactTypeWeb, notification::NotificationWeb};
 
 #[derive(Tsify, Debug, Serialize)]
 #[tsify(into_wasm_abi)]
@@ -131,10 +128,10 @@ pub struct BillCombinedBitcoinKeyWeb {
     pub private_key: String,
 }
 
-impl IntoWeb<BillCombinedBitcoinKeyWeb> for BillCombinedBitcoinKey {
-    fn into_web(self) -> BillCombinedBitcoinKeyWeb {
+impl From<BillCombinedBitcoinKey> for BillCombinedBitcoinKeyWeb {
+    fn from(val: BillCombinedBitcoinKey) -> Self {
         BillCombinedBitcoinKeyWeb {
-            private_key: self.private_key,
+            private_key: val.private_key,
         }
     }
 }
@@ -148,8 +145,8 @@ pub enum BillsFilterRoleWeb {
     Contingent,
 }
 
-impl FromWeb<BillsFilterRoleWeb> for BillsFilterRole {
-    fn from_web(value: BillsFilterRoleWeb) -> Self {
+impl From<BillsFilterRoleWeb> for BillsFilterRole {
+    fn from(value: BillsFilterRoleWeb) -> Self {
         match value {
             BillsFilterRoleWeb::All => BillsFilterRole::All,
             BillsFilterRoleWeb::Payer => BillsFilterRole::Payer,
@@ -168,13 +165,13 @@ pub struct PastEndorseeWeb {
     pub signing_address: Option<PostalAddressWeb>,
 }
 
-impl IntoWeb<PastEndorseeWeb> for PastEndorsee {
-    fn into_web(self) -> PastEndorseeWeb {
+impl From<PastEndorsee> for PastEndorseeWeb {
+    fn from(val: PastEndorsee) -> Self {
         PastEndorseeWeb {
-            pay_to_the_order_of: self.pay_to_the_order_of.into_web(),
-            signed: self.signed.into_web(),
-            signing_timestamp: self.signing_timestamp,
-            signing_address: self.signing_address.map(|s| s.into_web()),
+            pay_to_the_order_of: val.pay_to_the_order_of.into(),
+            signed: val.signed.into(),
+            signing_timestamp: val.signing_timestamp,
+            signing_address: val.signing_address.map(|s| s.into()),
         }
     }
 }
@@ -186,11 +183,11 @@ pub struct LightSignedByWeb {
     pub signatory: Option<LightBillIdentParticipantWeb>,
 }
 
-impl IntoWeb<LightSignedByWeb> for LightSignedBy {
-    fn into_web(self) -> LightSignedByWeb {
+impl From<LightSignedBy> for LightSignedByWeb {
+    fn from(val: LightSignedBy) -> Self {
         LightSignedByWeb {
-            data: self.data.into_web(),
-            signatory: self.signatory.map(|s| s.into_web()),
+            data: val.data.into(),
+            signatory: val.signatory.map(|s| s.into()),
         }
     }
 }
@@ -204,13 +201,13 @@ pub struct EndorsementWeb {
     pub signing_address: Option<PostalAddressWeb>,
 }
 
-impl IntoWeb<EndorsementWeb> for Endorsement {
-    fn into_web(self) -> EndorsementWeb {
+impl From<Endorsement> for EndorsementWeb {
+    fn from(val: Endorsement) -> Self {
         EndorsementWeb {
-            pay_to_the_order_of: self.pay_to_the_order_of.into_web(),
-            signed: self.signed.into_web(),
-            signing_timestamp: self.signing_timestamp,
-            signing_address: self.signing_address.map(|s| s.into_web()),
+            pay_to_the_order_of: val.pay_to_the_order_of.into(),
+            signed: val.signed.into(),
+            signing_timestamp: val.signing_timestamp,
+            signing_address: val.signing_address.map(|s| s.into()),
         }
     }
 }
@@ -275,12 +272,12 @@ pub enum PastPaymentResultWeb {
     Recourse(PastPaymentDataRecourseWeb),
 }
 
-impl IntoWeb<PastPaymentResultWeb> for PastPaymentResult {
-    fn into_web(self) -> PastPaymentResultWeb {
-        match self {
-            PastPaymentResult::Sell(state) => PastPaymentResultWeb::Sell(state.into_web()),
-            PastPaymentResult::Payment(state) => PastPaymentResultWeb::Payment(state.into_web()),
-            PastPaymentResult::Recourse(state) => PastPaymentResultWeb::Recourse(state.into_web()),
+impl From<PastPaymentResult> for PastPaymentResultWeb {
+    fn from(val: PastPaymentResult) -> Self {
+        match val {
+            PastPaymentResult::Sell(state) => PastPaymentResultWeb::Sell(state.into()),
+            PastPaymentResult::Payment(state) => PastPaymentResultWeb::Payment(state.into()),
+            PastPaymentResult::Recourse(state) => PastPaymentResultWeb::Recourse(state.into()),
         }
     }
 }
@@ -293,9 +290,9 @@ pub enum PastPaymentStatusWeb {
     Expired(u64),
 }
 
-impl IntoWeb<PastPaymentStatusWeb> for PastPaymentStatus {
-    fn into_web(self) -> PastPaymentStatusWeb {
-        match self {
+impl From<PastPaymentStatus> for PastPaymentStatusWeb {
+    fn from(val: PastPaymentStatus) -> Self {
+        match val {
             PastPaymentStatus::Paid(ts) => PastPaymentStatusWeb::Paid(ts),
             PastPaymentStatus::Rejected(ts) => PastPaymentStatusWeb::Rejected(ts),
             PastPaymentStatus::Expired(ts) => PastPaymentStatusWeb::Expired(ts),
@@ -318,19 +315,19 @@ pub struct PastPaymentDataSellWeb {
     pub status: PastPaymentStatusWeb,
 }
 
-impl IntoWeb<PastPaymentDataSellWeb> for PastPaymentDataSell {
-    fn into_web(self) -> PastPaymentDataSellWeb {
+impl From<PastPaymentDataSell> for PastPaymentDataSellWeb {
+    fn from(val: PastPaymentDataSell) -> Self {
         PastPaymentDataSellWeb {
-            time_of_request: self.time_of_request,
-            buyer: self.buyer.into_web(),
-            seller: self.seller.into_web(),
-            currency: self.currency,
-            sum: self.sum,
-            link_to_pay: self.link_to_pay,
-            address_to_pay: self.address_to_pay,
-            private_key_to_spend: self.private_key_to_spend,
-            mempool_link_for_address_to_pay: self.mempool_link_for_address_to_pay,
-            status: self.status.into_web(),
+            time_of_request: val.time_of_request,
+            buyer: val.buyer.into(),
+            seller: val.seller.into(),
+            currency: val.currency,
+            sum: val.sum,
+            link_to_pay: val.link_to_pay,
+            address_to_pay: val.address_to_pay,
+            private_key_to_spend: val.private_key_to_spend,
+            mempool_link_for_address_to_pay: val.mempool_link_for_address_to_pay,
+            status: val.status.into(),
         }
     }
 }
@@ -349,19 +346,19 @@ pub struct PastPaymentDataPaymentWeb {
     pub mempool_link_for_address_to_pay: String,
     pub status: PastPaymentStatusWeb,
 }
-impl IntoWeb<PastPaymentDataPaymentWeb> for PastPaymentDataPayment {
-    fn into_web(self) -> PastPaymentDataPaymentWeb {
+impl From<PastPaymentDataPayment> for PastPaymentDataPaymentWeb {
+    fn from(val: PastPaymentDataPayment) -> Self {
         PastPaymentDataPaymentWeb {
-            time_of_request: self.time_of_request,
-            payer: self.payer.into_web(),
-            payee: self.payee.into_web(),
-            currency: self.currency,
-            sum: self.sum,
-            link_to_pay: self.link_to_pay,
-            address_to_pay: self.address_to_pay,
-            private_key_to_spend: self.private_key_to_spend,
-            mempool_link_for_address_to_pay: self.mempool_link_for_address_to_pay,
-            status: self.status.into_web(),
+            time_of_request: val.time_of_request,
+            payer: val.payer.into(),
+            payee: val.payee.into(),
+            currency: val.currency,
+            sum: val.sum,
+            link_to_pay: val.link_to_pay,
+            address_to_pay: val.address_to_pay,
+            private_key_to_spend: val.private_key_to_spend,
+            mempool_link_for_address_to_pay: val.mempool_link_for_address_to_pay,
+            status: val.status.into(),
         }
     }
 }
@@ -381,19 +378,19 @@ pub struct PastPaymentDataRecourseWeb {
     pub status: PastPaymentStatusWeb,
 }
 
-impl IntoWeb<PastPaymentDataRecourseWeb> for PastPaymentDataRecourse {
-    fn into_web(self) -> PastPaymentDataRecourseWeb {
+impl From<PastPaymentDataRecourse> for PastPaymentDataRecourseWeb {
+    fn from(val: PastPaymentDataRecourse) -> Self {
         PastPaymentDataRecourseWeb {
-            time_of_request: self.time_of_request,
-            recourser: self.recourser.into_web(),
-            recoursee: self.recoursee.into_web(),
-            currency: self.currency,
-            sum: self.sum,
-            link_to_pay: self.link_to_pay,
-            address_to_pay: self.address_to_pay,
-            private_key_to_spend: self.private_key_to_spend,
-            mempool_link_for_address_to_pay: self.mempool_link_for_address_to_pay,
-            status: self.status.into_web(),
+            time_of_request: val.time_of_request,
+            recourser: val.recourser.into(),
+            recoursee: val.recoursee.into(),
+            currency: val.currency,
+            sum: val.sum,
+            link_to_pay: val.link_to_pay,
+            address_to_pay: val.address_to_pay,
+            private_key_to_spend: val.private_key_to_spend,
+            mempool_link_for_address_to_pay: val.mempool_link_for_address_to_pay,
+            status: val.status.into(),
         }
     }
 }
@@ -420,14 +417,14 @@ pub struct BitcreditBillWeb {
     pub current_waiting_state: Option<BillCurrentWaitingStateWeb>,
 }
 
-impl IntoWeb<BitcreditBillWeb> for BitcreditBillResult {
-    fn into_web(self) -> BitcreditBillWeb {
+impl From<BitcreditBillResult> for BitcreditBillWeb {
+    fn from(val: BitcreditBillResult) -> Self {
         BitcreditBillWeb {
-            id: self.id,
-            participants: self.participants.into_web(),
-            data: self.data.into_web(),
-            status: self.status.into_web(),
-            current_waiting_state: self.current_waiting_state.map(|cws| cws.into_web()),
+            id: val.id,
+            participants: val.participants.into(),
+            data: val.data.into(),
+            status: val.status.into(),
+            current_waiting_state: val.current_waiting_state.map(|cws| cws.into()),
         }
     }
 }
@@ -440,17 +437,15 @@ pub enum BillCurrentWaitingStateWeb {
     Recourse(BillWaitingForRecourseStateWeb),
 }
 
-impl IntoWeb<BillCurrentWaitingStateWeb> for BillCurrentWaitingState {
-    fn into_web(self) -> BillCurrentWaitingStateWeb {
-        match self {
-            BillCurrentWaitingState::Sell(state) => {
-                BillCurrentWaitingStateWeb::Sell(state.into_web())
-            }
+impl From<BillCurrentWaitingState> for BillCurrentWaitingStateWeb {
+    fn from(val: BillCurrentWaitingState) -> Self {
+        match val {
+            BillCurrentWaitingState::Sell(state) => BillCurrentWaitingStateWeb::Sell(state.into()),
             BillCurrentWaitingState::Payment(state) => {
-                BillCurrentWaitingStateWeb::Payment(state.into_web())
+                BillCurrentWaitingStateWeb::Payment(state.into())
             }
             BillCurrentWaitingState::Recourse(state) => {
-                BillCurrentWaitingStateWeb::Recourse(state.into_web())
+                BillCurrentWaitingStateWeb::Recourse(state.into())
             }
         }
     }
@@ -469,17 +464,17 @@ pub struct BillWaitingForSellStateWeb {
     pub mempool_link_for_address_to_pay: String,
 }
 
-impl IntoWeb<BillWaitingForSellStateWeb> for BillWaitingForSellState {
-    fn into_web(self) -> BillWaitingForSellStateWeb {
+impl From<BillWaitingForSellState> for BillWaitingForSellStateWeb {
+    fn from(val: BillWaitingForSellState) -> Self {
         BillWaitingForSellStateWeb {
-            time_of_request: self.time_of_request,
-            buyer: self.buyer.into_web(),
-            seller: self.seller.into_web(),
-            currency: self.currency,
-            sum: self.sum,
-            link_to_pay: self.link_to_pay,
-            address_to_pay: self.address_to_pay,
-            mempool_link_for_address_to_pay: self.mempool_link_for_address_to_pay,
+            time_of_request: val.time_of_request,
+            buyer: val.buyer.into(),
+            seller: val.seller.into(),
+            currency: val.currency,
+            sum: val.sum,
+            link_to_pay: val.link_to_pay,
+            address_to_pay: val.address_to_pay,
+            mempool_link_for_address_to_pay: val.mempool_link_for_address_to_pay,
         }
     }
 }
@@ -497,17 +492,17 @@ pub struct BillWaitingForPaymentStateWeb {
     pub mempool_link_for_address_to_pay: String,
 }
 
-impl IntoWeb<BillWaitingForPaymentStateWeb> for BillWaitingForPaymentState {
-    fn into_web(self) -> BillWaitingForPaymentStateWeb {
+impl From<BillWaitingForPaymentState> for BillWaitingForPaymentStateWeb {
+    fn from(val: BillWaitingForPaymentState) -> Self {
         BillWaitingForPaymentStateWeb {
-            time_of_request: self.time_of_request,
-            payer: self.payer.into_web(),
-            payee: self.payee.into_web(),
-            currency: self.currency,
-            sum: self.sum,
-            link_to_pay: self.link_to_pay,
-            address_to_pay: self.address_to_pay,
-            mempool_link_for_address_to_pay: self.mempool_link_for_address_to_pay,
+            time_of_request: val.time_of_request,
+            payer: val.payer.into(),
+            payee: val.payee.into(),
+            currency: val.currency,
+            sum: val.sum,
+            link_to_pay: val.link_to_pay,
+            address_to_pay: val.address_to_pay,
+            mempool_link_for_address_to_pay: val.mempool_link_for_address_to_pay,
         }
     }
 }
@@ -524,17 +519,17 @@ pub struct BillWaitingForRecourseStateWeb {
     pub address_to_pay: String,
     pub mempool_link_for_address_to_pay: String,
 }
-impl IntoWeb<BillWaitingForRecourseStateWeb> for BillWaitingForRecourseState {
-    fn into_web(self) -> BillWaitingForRecourseStateWeb {
+impl From<BillWaitingForRecourseState> for BillWaitingForRecourseStateWeb {
+    fn from(val: BillWaitingForRecourseState) -> Self {
         BillWaitingForRecourseStateWeb {
-            time_of_request: self.time_of_request,
-            recourser: self.recourser.into_web(),
-            recoursee: self.recoursee.into_web(),
-            currency: self.currency,
-            sum: self.sum,
-            link_to_pay: self.link_to_pay,
-            address_to_pay: self.address_to_pay,
-            mempool_link_for_address_to_pay: self.mempool_link_for_address_to_pay,
+            time_of_request: val.time_of_request,
+            recourser: val.recourser.into(),
+            recoursee: val.recoursee.into(),
+            currency: val.currency,
+            sum: val.sum,
+            link_to_pay: val.link_to_pay,
+            address_to_pay: val.address_to_pay,
+            mempool_link_for_address_to_pay: val.mempool_link_for_address_to_pay,
         }
     }
 }
@@ -550,15 +545,15 @@ pub struct BillStatusWeb {
     pub has_requested_funds: bool,
 }
 
-impl IntoWeb<BillStatusWeb> for BillStatus {
-    fn into_web(self) -> BillStatusWeb {
+impl From<BillStatus> for BillStatusWeb {
+    fn from(val: BillStatus) -> Self {
         BillStatusWeb {
-            acceptance: self.acceptance.into_web(),
-            payment: self.payment.into_web(),
-            sell: self.sell.into_web(),
-            recourse: self.recourse.into_web(),
-            redeemed_funds_available: self.redeemed_funds_available,
-            has_requested_funds: self.has_requested_funds,
+            acceptance: val.acceptance.into(),
+            payment: val.payment.into(),
+            sell: val.sell.into(),
+            recourse: val.recourse.into(),
+            redeemed_funds_available: val.redeemed_funds_available,
+            has_requested_funds: val.has_requested_funds,
         }
     }
 }
@@ -573,14 +568,14 @@ pub struct BillAcceptanceStatusWeb {
     pub rejected_to_accept: bool,
 }
 
-impl IntoWeb<BillAcceptanceStatusWeb> for BillAcceptanceStatus {
-    fn into_web(self) -> BillAcceptanceStatusWeb {
+impl From<BillAcceptanceStatus> for BillAcceptanceStatusWeb {
+    fn from(val: BillAcceptanceStatus) -> Self {
         BillAcceptanceStatusWeb {
-            time_of_request_to_accept: self.time_of_request_to_accept,
-            requested_to_accept: self.requested_to_accept,
-            accepted: self.accepted,
-            request_to_accept_timed_out: self.request_to_accept_timed_out,
-            rejected_to_accept: self.rejected_to_accept,
+            time_of_request_to_accept: val.time_of_request_to_accept,
+            requested_to_accept: val.requested_to_accept,
+            accepted: val.accepted,
+            request_to_accept_timed_out: val.request_to_accept_timed_out,
+            rejected_to_accept: val.rejected_to_accept,
         }
     }
 }
@@ -594,14 +589,14 @@ pub struct BillPaymentStatusWeb {
     pub request_to_pay_timed_out: bool,
     pub rejected_to_pay: bool,
 }
-impl IntoWeb<BillPaymentStatusWeb> for BillPaymentStatus {
-    fn into_web(self) -> BillPaymentStatusWeb {
+impl From<BillPaymentStatus> for BillPaymentStatusWeb {
+    fn from(val: BillPaymentStatus) -> Self {
         BillPaymentStatusWeb {
-            time_of_request_to_pay: self.time_of_request_to_pay,
-            requested_to_pay: self.requested_to_pay,
-            paid: self.paid,
-            request_to_pay_timed_out: self.request_to_pay_timed_out,
-            rejected_to_pay: self.rejected_to_pay,
+            time_of_request_to_pay: val.time_of_request_to_pay,
+            requested_to_pay: val.requested_to_pay,
+            paid: val.paid,
+            request_to_pay_timed_out: val.request_to_pay_timed_out,
+            rejected_to_pay: val.rejected_to_pay,
         }
     }
 }
@@ -615,14 +610,14 @@ pub struct BillSellStatusWeb {
     pub offer_to_sell_timed_out: bool,
     pub rejected_offer_to_sell: bool,
 }
-impl IntoWeb<BillSellStatusWeb> for BillSellStatus {
-    fn into_web(self) -> BillSellStatusWeb {
+impl From<BillSellStatus> for BillSellStatusWeb {
+    fn from(val: BillSellStatus) -> Self {
         BillSellStatusWeb {
-            time_of_last_offer_to_sell: self.time_of_last_offer_to_sell,
-            sold: self.sold,
-            offered_to_sell: self.offered_to_sell,
-            offer_to_sell_timed_out: self.offer_to_sell_timed_out,
-            rejected_offer_to_sell: self.rejected_offer_to_sell,
+            time_of_last_offer_to_sell: val.time_of_last_offer_to_sell,
+            sold: val.sold,
+            offered_to_sell: val.offered_to_sell,
+            offer_to_sell_timed_out: val.offer_to_sell_timed_out,
+            rejected_offer_to_sell: val.rejected_offer_to_sell,
         }
     }
 }
@@ -637,14 +632,14 @@ pub struct BillRecourseStatusWeb {
     pub rejected_request_to_recourse: bool,
 }
 
-impl IntoWeb<BillRecourseStatusWeb> for BillRecourseStatus {
-    fn into_web(self) -> BillRecourseStatusWeb {
+impl From<BillRecourseStatus> for BillRecourseStatusWeb {
+    fn from(val: BillRecourseStatus) -> Self {
         BillRecourseStatusWeb {
-            time_of_last_request_to_recourse: self.time_of_last_request_to_recourse,
-            recoursed: self.recoursed,
-            requested_to_recourse: self.requested_to_recourse,
-            request_to_recourse_timed_out: self.request_to_recourse_timed_out,
-            rejected_request_to_recourse: self.rejected_request_to_recourse,
+            time_of_last_request_to_recourse: val.time_of_last_request_to_recourse,
+            recoursed: val.recoursed,
+            requested_to_recourse: val.requested_to_recourse,
+            request_to_recourse_timed_out: val.request_to_recourse_timed_out,
+            rejected_request_to_recourse: val.rejected_request_to_recourse,
         }
     }
 }
@@ -667,22 +662,22 @@ pub struct BillDataWeb {
     pub active_notification: Option<NotificationWeb>,
 }
 
-impl IntoWeb<BillDataWeb> for BillData {
-    fn into_web(self) -> BillDataWeb {
+impl From<BillData> for BillDataWeb {
+    fn from(val: BillData) -> Self {
         BillDataWeb {
-            language: self.language,
-            time_of_drawing: self.time_of_drawing,
-            issue_date: self.issue_date,
-            time_of_maturity: self.time_of_maturity,
-            maturity_date: self.maturity_date,
-            country_of_issuing: self.country_of_issuing,
-            city_of_issuing: self.city_of_issuing,
-            country_of_payment: self.country_of_payment,
-            city_of_payment: self.city_of_payment,
-            currency: self.currency,
-            sum: self.sum,
-            files: self.files.into_iter().map(|f| f.into_web()).collect(),
-            active_notification: self.active_notification.map(|an| an.into_web()),
+            language: val.language,
+            time_of_drawing: val.time_of_drawing,
+            issue_date: val.issue_date,
+            time_of_maturity: val.time_of_maturity,
+            maturity_date: val.maturity_date,
+            country_of_issuing: val.country_of_issuing,
+            city_of_issuing: val.city_of_issuing,
+            country_of_payment: val.country_of_payment,
+            city_of_payment: val.city_of_payment,
+            currency: val.currency,
+            sum: val.sum,
+            files: val.files.into_iter().map(|f| f.into()).collect(),
+            active_notification: val.active_notification.map(|an| an.into()),
         }
     }
 }
@@ -698,15 +693,15 @@ pub struct BillParticipantsWeb {
     pub all_participant_node_ids: Vec<String>,
 }
 
-impl IntoWeb<BillParticipantsWeb> for BillParticipants {
-    fn into_web(self) -> BillParticipantsWeb {
+impl From<BillParticipants> for BillParticipantsWeb {
+    fn from(val: BillParticipants) -> Self {
         BillParticipantsWeb {
-            drawee: self.drawee.into_web(),
-            drawer: self.drawer.into_web(),
-            payee: self.payee.into_web(),
-            endorsee: self.endorsee.map(|e| e.into_web()),
-            endorsements_count: self.endorsements_count,
-            all_participant_node_ids: self.all_participant_node_ids,
+            drawee: val.drawee.into(),
+            drawer: val.drawer.into(),
+            payee: val.payee.into(),
+            endorsee: val.endorsee.map(|e| e.into()),
+            endorsements_count: val.endorsements_count,
+            all_participant_node_ids: val.all_participant_node_ids,
         }
     }
 }
@@ -727,20 +722,20 @@ pub struct LightBitcreditBillWeb {
     pub time_of_maturity: u64,
 }
 
-impl IntoWeb<LightBitcreditBillWeb> for LightBitcreditBillResult {
-    fn into_web(self) -> LightBitcreditBillWeb {
+impl From<LightBitcreditBillResult> for LightBitcreditBillWeb {
+    fn from(val: LightBitcreditBillResult) -> Self {
         LightBitcreditBillWeb {
-            id: self.id,
-            drawee: self.drawee.into_web(),
-            drawer: self.drawer.into_web(),
-            payee: self.payee.into_web(),
-            endorsee: self.endorsee.map(|e| e.into_web()),
-            active_notification: self.active_notification.map(|n| n.into_web()),
-            sum: self.sum,
-            currency: self.currency,
-            issue_date: self.issue_date,
-            time_of_drawing: self.time_of_drawing,
-            time_of_maturity: self.time_of_maturity,
+            id: val.id,
+            drawee: val.drawee.into(),
+            drawer: val.drawer.into(),
+            payee: val.payee.into(),
+            endorsee: val.endorsee.map(|e| e.into()),
+            active_notification: val.active_notification.map(|n| n.into()),
+            sum: val.sum,
+            currency: val.currency,
+            issue_date: val.issue_date,
+            time_of_drawing: val.time_of_drawing,
+            time_of_maturity: val.time_of_maturity,
         }
     }
 }
@@ -752,11 +747,11 @@ pub enum BillParticipantWeb {
     Ident(BillIdentParticipantWeb),
 }
 
-impl IntoWeb<BillParticipantWeb> for BillParticipant {
-    fn into_web(self) -> BillParticipantWeb {
-        match self {
-            BillParticipant::Ident(data) => BillParticipantWeb::Ident(data.into_web()),
-            BillParticipant::Anon(data) => BillParticipantWeb::Anon(data.into_web()),
+impl From<BillParticipant> for BillParticipantWeb {
+    fn from(val: BillParticipant) -> Self {
+        match val {
+            BillParticipant::Ident(data) => BillParticipantWeb::Ident(data.into()),
+            BillParticipant::Anon(data) => BillParticipantWeb::Anon(data.into()),
         }
     }
 }
@@ -769,12 +764,12 @@ pub struct BillAnonParticipantWeb {
     pub nostr_relays: Vec<String>,
 }
 
-impl IntoWeb<BillAnonParticipantWeb> for BillAnonParticipant {
-    fn into_web(self) -> BillAnonParticipantWeb {
+impl From<BillAnonParticipant> for BillAnonParticipantWeb {
+    fn from(val: BillAnonParticipant) -> Self {
         BillAnonParticipantWeb {
-            node_id: self.node_id,
-            email: self.email,
-            nostr_relays: self.nostr_relays,
+            node_id: val.node_id,
+            email: val.email,
+            nostr_relays: val.nostr_relays,
         }
     }
 }
@@ -790,15 +785,15 @@ pub struct BillIdentParticipantWeb {
     pub nostr_relays: Vec<String>,
 }
 
-impl IntoWeb<BillIdentParticipantWeb> for BillIdentParticipant {
-    fn into_web(self) -> BillIdentParticipantWeb {
+impl From<BillIdentParticipant> for BillIdentParticipantWeb {
+    fn from(val: BillIdentParticipant) -> Self {
         BillIdentParticipantWeb {
-            t: self.t.into_web(),
-            name: self.name,
-            node_id: self.node_id,
-            postal_address: self.postal_address.into_web(),
-            email: self.email,
-            nostr_relays: self.nostr_relays,
+            t: val.t.into(),
+            name: val.name,
+            node_id: val.node_id,
+            postal_address: val.postal_address.into(),
+            email: val.email,
+            nostr_relays: val.nostr_relays,
         }
     }
 }
@@ -812,13 +807,13 @@ pub struct LightBillIdentParticipantWithAddressWeb {
     pub postal_address: PostalAddressWeb,
 }
 
-impl IntoWeb<LightBillIdentParticipantWithAddressWeb> for LightBillIdentParticipantWithAddress {
-    fn into_web(self) -> LightBillIdentParticipantWithAddressWeb {
+impl From<LightBillIdentParticipantWithAddress> for LightBillIdentParticipantWithAddressWeb {
+    fn from(val: LightBillIdentParticipantWithAddress) -> Self {
         LightBillIdentParticipantWithAddressWeb {
-            t: self.t.into_web(),
-            name: self.name,
-            node_id: self.node_id,
-            postal_address: self.postal_address.into_web(),
+            t: val.t.into(),
+            name: val.name,
+            node_id: val.node_id,
+            postal_address: val.postal_address.into(),
         }
     }
 }
@@ -830,11 +825,11 @@ pub enum LightBillParticipantWeb {
     Ident(LightBillIdentParticipantWeb),
 }
 
-impl IntoWeb<LightBillParticipantWeb> for LightBillParticipant {
-    fn into_web(self) -> LightBillParticipantWeb {
-        match self {
-            LightBillParticipant::Ident(data) => LightBillParticipantWeb::Ident(data.into_web()),
-            LightBillParticipant::Anon(data) => LightBillParticipantWeb::Anon(data.into_web()),
+impl From<LightBillParticipant> for LightBillParticipantWeb {
+    fn from(val: LightBillParticipant) -> Self {
+        match val {
+            LightBillParticipant::Ident(data) => LightBillParticipantWeb::Ident(data.into()),
+            LightBillParticipant::Anon(data) => LightBillParticipantWeb::Anon(data.into()),
         }
     }
 }
@@ -845,10 +840,10 @@ pub struct LightBillAnonParticipantWeb {
     pub node_id: String,
 }
 
-impl IntoWeb<LightBillAnonParticipantWeb> for LightBillAnonParticipant {
-    fn into_web(self) -> LightBillAnonParticipantWeb {
+impl From<LightBillAnonParticipant> for LightBillAnonParticipantWeb {
+    fn from(val: LightBillAnonParticipant) -> Self {
         LightBillAnonParticipantWeb {
-            node_id: self.node_id,
+            node_id: val.node_id,
         }
     }
 }
@@ -861,12 +856,12 @@ pub struct LightBillIdentParticipantWeb {
     pub node_id: String,
 }
 
-impl IntoWeb<LightBillIdentParticipantWeb> for LightBillIdentParticipant {
-    fn into_web(self) -> LightBillIdentParticipantWeb {
+impl From<LightBillIdentParticipant> for LightBillIdentParticipantWeb {
+    fn from(val: LightBillIdentParticipant) -> Self {
         LightBillIdentParticipantWeb {
-            t: self.t.into_web(),
-            name: self.name,
-            node_id: self.node_id,
+            t: val.t.into(),
+            name: val.name,
+            node_id: val.node_id,
         }
     }
 }